+++ services/interfaces/jd_adc.h
@@ -5,7 +5,7 @@
 #define __JD_ADC_H
 
 void adc_init_random(void);
-uint16_t adc_read_temp(void);
+int32_t adc_read_temp(void);
 bool adc_can_read_pin(uint8_t pin);
 
 void adc_prep_read_pin(uint8_t pin);
